little fix

+++ webexample/forms.py
@@ -3,6 +3,6 @@ from django import forms
 
 class PriceForm(forms.Form):
     order = forms.CharField()
-    price_usa = forms.IntegerField()
-    price_rus = forms.FloatField()
+    price_usa = forms.DecimalField()
+    price_rus = forms.DecimalField()
     delivery_date = forms.DateTimeField()
